requirements_test.txt
pytest>=7.0
pytest-asyncio>=0.21
pytest-homeassistant-custom-component>=2024.1.0
